requirements.txt
discord.py
tbapy
airtable